+++ apps/docs/CHANGELOG.md
@@ -6,5 +6,5 @@
 
 - changelog init
 - Updated dependencies
-  - @awsm/fonts@0.0.1
-  - @awsm/tokens@0.0.1
+  - @ds-starter/fonts@0.0.1
+  - @ds-starter/tokens@0.0.1